move enginecache func to new versioncheck package
